Revert "Add experimental setting to define where files should be copied for markdown (#169454)" (#169615)
This reverts commit 1d6b47a005f6597a82c18587718eee99f7d5acda.
