+++ extensions/markdown-language-features/package.json
@@ -571,13 +571,6 @@
           "default": false,
           "description": "%configuration.markdown.occurrencesHighlight.enabled%",
           "scope": "resource"
-        },
-        "markdown.experimental.copyFiles.destination": {
-          "type": "object",
-          "markdownDescription": "%configuration.markdown.copyFiles.destination%",
-          "additionalProperties": {
-            "type": "string"
-          }
         }
       }
     },
