+++ extensions/markdown-language-features/src/extension.shared.ts
@@ -7,9 +7,9 @@ import * as vscode from 'vscode';
 import { MdLanguageClient } from './client/client';
 import { CommandManager } from './commandManager';
 import { registerMarkdownCommands } from './commands/index';
-import { registerPasteSupport } from './languageFeatures/copyFiles/copyPaste';
+import { registerPasteSupport } from './languageFeatures/copyPaste';
 import { registerDiagnosticSupport } from './languageFeatures/diagnostics';
-import { registerDropIntoEditorSupport } from './languageFeatures/copyFiles/dropIntoEditor';
+import { registerDropIntoEditorSupport } from './languageFeatures/dropIntoEditor';
 import { registerFindFileReferenceSupport } from './languageFeatures/fileReferences';
 import { registerUpdateLinksOnRename } from './languageFeatures/linkUpdater';
 import { ILogger } from './logging';
