+++ extensions/markdown-language-features/src/languageFeatures/copyPaste.ts
@@ -3,9 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as path from 'path';
 import * as vscode from 'vscode';
-import { Schemes } from '../../util/schemes';
-import { getNewFileName } from './copyFiles';
+import { Utils } from 'vscode-uri';
+import { Schemes } from '../util/schemes';
 import { createUriListSnippet, tryGetUriListSnippet } from './dropIntoEditor';
 
 const supportedImageMimes = new Set([
@@ -58,7 +59,7 @@ class PasteEditProvider implements vscode.DocumentPasteEditProvider {
 			}
 		}
 
-		const uri = await getNewFileName(document, file);
+		const uri = await this._getNewFileName(document, file);
 		if (token.isCancellationRequested) {
 			return;
 		}
@@ -75,6 +76,23 @@ class PasteEditProvider implements vscode.DocumentPasteEditProvider {
 		const pasteEdit = new vscode.DocumentPasteEdit(snippet);
 		pasteEdit.additionalEdit = workspaceEdit;
 		return pasteEdit;
+	}
+
+	private async _getNewFileName(document: vscode.TextDocument, file: vscode.DataTransferFile): Promise<vscode.Uri> {
+		const root = Utils.dirname(document.uri);
+
+		const ext = path.extname(file.name);
+		const baseName = path.basename(file.name, ext);
+		for (let i = 0; ; ++i) {
+			const name = i === 0 ? baseName : `${baseName}-${i}`;
+			const uri = vscode.Uri.joinPath(root, `${name}${ext}`);
+			try {
+				await vscode.workspace.fs.stat(uri);
+			} catch {
+				// Does not exist
+				return uri;
+			}
+		}
 	}
 }
